https://leetcode.com/problems/h-index/ https://leetcode.com/problems/trapping-rain-water/ https://leetcode.com/problems/rotate-array
