+++ Problem274.java
@@ -1,72 +1,52 @@
 import java.util.Arrays;
 
 public class Problem274 {
-    //TC=O(n)
-//SC=O(1)
-    public int trap1(int[] height) {
-        int n=height.length;
-        int max=0;
-        int maxHeightIdx=-1;
-        int maxArea=0;
+    public int hIndex1(int[] citations) {
+        //TC=O(nlogn)
+        //SC=O(1)
+        if(citations==null || citations.length==0)
+            return 0;
+        Arrays.sort(citations);
+        int low=0;
+        int n=citations.length;
+        int high=n-1;
+        while(low<=high){
+            int mid=low+(high-low)/2;
+            int diff=n-mid;
+            if(citations[mid]==diff)
+                return diff;
+            else if(citations[mid]>diff)
+                high=mid-1;
+            else
+                low=mid+1;
+        }
 
+        return n-low;
+    }
+    public int hIndex(int[] citations) {
+        //TC=O(n)
+        //SC=O(n)
+        int n=citations.length;
+        if(citations==null || citations.length==0)
+            return 0;
+        int[] bucketList=new int[n+1];
         for(int i=0;i<n;i++){
-            if(height[i]>max){
-                max=height[i];
-                maxHeightIdx=i;
-            }
+            if(citations[i]>=n)
+                bucketList[n]++;
+            else
+                bucketList[citations[i]]++;
         }
-        int lw=0;
-        for(int j=0;j<maxHeightIdx;j++){
-            if(height[j]<lw){
-                maxArea+=lw-height[j];
-            }
-            else{
-                lw=height[j];
-            }
+        int sum=0;
+        for(int i=n;i>=0;i--){
+            sum+=bucketList[i];
+            if(sum>=i)
+                return i;
         }
-        int rw=0;
-        for(int j=n-1;j>maxHeightIdx;j--){
-            if(height[j]<rw){
-                maxArea+=rw-height[j];
-            }
-            else{
-                rw=height[j];
-            }
-        }
-        return maxArea;
+        return -1;
     }
-    public int trap(int[] height) {
-        int rightPointer=height.length-1;
-        int lw=0;
-        int rw=0;
-        int leftPointer=0;
-        int maxArea=0;
-        while(leftPointer<=rightPointer){
-            if(lw<=rw){
-                if(height[leftPointer]<lw){
-                    maxArea+=lw-height[leftPointer];
-                }
-                else{
-                    lw=height[leftPointer];
-                }
-                leftPointer++;
-            }
-            else{
-                if(height[rightPointer]<rw){
-                    maxArea+=rw-height[rightPointer];
-                }
-                else{
-                    rw=height[rightPointer];
-                }
-                rightPointer--;
-            }
-        }
-        return maxArea;
-    }
-
     public static void main(String[] args) {
         Problem274 problem=new Problem274();
-        System.out.println(problem.trap(new int[]{0,1,0,2,1,0,1,3,2,1,2,1}));
-        System.out.println(problem.trap1(new int[]{0,1,0,2,1,0,1,3,2,1,2,1}));
+        System.out.println(problem.hIndex(new int[]{3,0,6,1,5}));
+        System.out.println(problem.hIndex1(new int[]{3,0,6,1,5}));
     }
 }
